fix: Physically correct lighting toggle (#7)
* refactor: Move normal map methods

- Normal map and AO functions moved to own file
- Further optimized applyPbrMaterial function
- Optimized channel assignment setup/teardown

* fix: Correct lighting button

- Fixed physically correct lighting toggle
- DRY code
